requirements.txt
streamlit==1.29.0
openai>=2.9.0
python-dotenv==1.0.0
plotly==5.18.0
pandas==2.1.4
